requirements.txt
streamlit==0.82.0
tensorflow==2.5.0
pillow==8.2.0
